cr plus types
