+++ tensorboard/webapp/metrics/BUILD
@@ -26,20 +26,20 @@ tf_ng_module(
 )
 
 tf_ts_library(
-    name = "public_types",
+    name = "types",
     srcs = [
-        "public_types.ts",
+        "types.ts",
     ],
     deps = [
-        ":types",
+        ":internal_types",
         "//tensorboard/webapp/metrics/store:types",
     ],
 )
 
 tf_ts_library(
-    name = "types",
+    name = "internal_types",
     srcs = [
-        "types.ts",
+        "internal_types.ts",
     ],
     visibility = ["//tensorboard/webapp/metrics:__subpackages__"],
     deps = [
