+++ tensorboard/webapp/metrics/store/BUILD
@@ -17,7 +17,7 @@ tf_ts_library(
         "//tensorboard/webapp/app_routing:types",
         "//tensorboard/webapp/app_routing/actions",
         "//tensorboard/webapp/core/actions",
-        "//tensorboard/webapp/metrics:types",
+        "//tensorboard/webapp/metrics:internal_types",
         "//tensorboard/webapp/metrics/actions",
         "//tensorboard/webapp/metrics/data_source",
         "//tensorboard/webapp/types",
@@ -36,7 +36,7 @@ tf_ts_library(
     visibility = ["//tensorboard/webapp/metrics/store:__subpackages__"],
     deps = [
         ":types",
-        "//tensorboard/webapp/metrics:types",
+        "//tensorboard/webapp/metrics:internal_types",
         "//tensorboard/webapp/metrics/data_source",
         "//tensorboard/webapp/types",
     ],
@@ -49,7 +49,7 @@ tf_ts_library(
     ],
     deps = [
         "//tensorboard/webapp/app_routing:route_contexted_reducer_helper",
-        "//tensorboard/webapp/metrics:types",
+        "//tensorboard/webapp/metrics:internal_types",
         "//tensorboard/webapp/metrics/data_source",
         "//tensorboard/webapp/types",
     ],
@@ -71,8 +71,8 @@ tf_ts_library(
         "//tensorboard/webapp/app_routing:types",
         "//tensorboard/webapp/app_routing/actions",
         "//tensorboard/webapp/core/actions",
+        "//tensorboard/webapp/metrics:internal_types",
         "//tensorboard/webapp/metrics:test_lib",
-        "//tensorboard/webapp/metrics:types",
         "//tensorboard/webapp/metrics/actions",
         "//tensorboard/webapp/metrics/data_source",
         "//tensorboard/webapp/types",
