+++ tensorboard/webapp/metrics/store/metrics_reducers_test.ts
@@ -12,18 +12,24 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-import * as coreActions from '../../core/actions';
-import {DataLoadState} from '../../types/data';
-
 import * as routingActions from '../../app_routing/actions';
 import {buildRoute} from '../../app_routing/testing';
 import {RouteKind} from '../../app_routing/types';
+import * as coreActions from '../../core/actions';
+import {DataLoadState} from '../../types/data';
 import * as actions from '../actions';
 import {
   PluginType,
   ScalarStepDatum,
   TagMetadata as DataSourceTagMetadata,
 } from '../data_source';
+import {
+  CardId,
+  CardMetadata,
+  HistogramMode,
+  TooltipSort,
+  XAxisType,
+} from '../internal_types';
 import {
   buildDataSourceTagMetadata,
   buildMetricsSettingsState,
@@ -36,14 +42,6 @@ import {
   createScalarStepData,
   createTimeSeriesData,
 } from '../testing';
-import {
-  CardId,
-  CardMetadata,
-  HistogramMode,
-  TooltipSort,
-  XAxisType,
-} from '../types';
-
 import {reducers} from './metrics_reducers';
 import {getCardId, getPinnedCardId} from './metrics_store_internal_utils';
 import {
@@ -1658,6 +1656,19 @@ describe('metrics reducers', () => {
       const nextState = reducers(beforeState, action);
 
       expect(nextState.settings.scalarSmoothing).toBe(0.999);
+    });
+
+    it('rounds to the 3 significant digits to prevent weird numbers', () => {
+      const beforeState = buildMetricsState({
+        settings: buildMetricsSettingsState({scalarSmoothing: 0.3}),
+      });
+      const action = routingActions.stateRehydratedFromUrl({
+        routeKind: RouteKind.EXPERIMENT,
+        partialState: {metrics: {pinnedCards: [], smoothing: 0.2318421}},
+      });
+      const nextState = reducers(beforeState, action);
+
+      expect(nextState.settings.scalarSmoothing).toBe(0.232);
     });
   });
 });
