+++ tensorboard/webapp/metrics/store/metrics_store_internal_utils.ts
@@ -19,7 +19,12 @@ limitations under the License.
 import {DataLoadState} from '../../types/data';
 
 import {isSampledPlugin, PluginType, SampledPluginType} from '../data_source';
-import {CardId, CardMetadata, CardUniqueInfo, NonPinnedCardId} from '../types';
+import {
+  CardId,
+  CardMetadata,
+  CardUniqueInfo,
+  NonPinnedCardId,
+} from '../internal_types';
 
 import {
   CardMetadataMap,
