+++ tensorboard/webapp/metrics/types.ts
@@ -1,4 +1,4 @@
-/* Copyright 2020 The TensorFlow Authors. All Rights Reserved.
+/* Copyright 2021 The TensorFlow Authors. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -12,69 +12,5 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-import {HistogramMode} from '../widgets/histogram/histogram_types';
-import {TooltipSortingMethod} from '../widgets/line_chart/polymer_interop_types';
-
-import {PluginType} from './data_source/types';
-
-export {HistogramMode};
-export {TooltipSortingMethod as TooltipSort};
-
-export enum XAxisType {
-  STEP,
-  RELATIVE,
-  WALL_TIME,
-}
-
-export interface CardMetadata {
-  plugin: PluginType;
-  tag: string;
-  sample?: number;
-
-  /**
-   * A `null` runId indicates all runs.
-   */
-  runId: string | null;
-}
-
-/**
- * Identifier for a card which, itself, is not a pinned copy of another card.
- * Views may treat a non-pinned card as having "truthy pin state" if another
- * `PinnedCardId` is associated to it.
- */
-export type NonPinnedCardId = string;
-
-export type PinnedCardId = string;
-
-/**
- * A unique identifier to a specific card instance in the UI. This is an opaque
- * ID, meaning that consumers should never peer into/parse it and never assume
- * that it will always be a string.
- */
-export type CardId = NonPinnedCardId | PinnedCardId;
-
-export type CardIdWithMetadata = CardMetadata & {
-  cardId: CardId;
-};
-
-/**
- * The most minimal representation of a card that uniquely identifies it across
- * a browser session. This information may be persisted in storage, retrieved,
- * and used to match against an existing card with the same metadata.
- */
-export interface CardUniqueInfo {
-  plugin: string;
-  tag: string;
-  runId?: string;
-  sample?: number;
-}
-
-/**
- * The metrics-related state created by deserializing a URL.
- */
-export interface URLDeserializedState {
-  metrics: {
-    pinnedCards: CardUniqueInfo[];
-    smoothing: number | null;
-  };
-}
+export * from './internal_types';
+export {METRICS_SETTINGS_DEFAULT} from './store/metrics_types';
