+++ tensorboard/webapp/routes/BUILD
@@ -24,7 +24,7 @@ tf_ts_library(
         "core_deeplink_provider_types.ts",
     ],
     deps = [
-        "//tensorboard/webapp/metrics:public_types",
+        "//tensorboard/webapp/metrics:types",
     ],
 )
 
@@ -40,7 +40,7 @@ tf_ts_library(
         "//tensorboard/webapp/app_routing:deep_link_provider",
         "//tensorboard/webapp/app_routing:route_config",
         "//tensorboard/webapp/app_routing:types",
-        "//tensorboard/webapp/metrics:public_types",
+        "//tensorboard/webapp/metrics:types",
         "//tensorboard/webapp/metrics/data_source:types",
         "//tensorboard/webapp/tb_wrapper",
         "//tensorboard/webapp/webapp_data_source:feature_flag_types",
@@ -64,8 +64,8 @@ tf_ts_library(
         "//tensorboard/webapp/angular:expect_ngrx_store_testing",
         "//tensorboard/webapp/app_routing:deep_link_provider",
         "//tensorboard/webapp/app_routing:types",
-        "//tensorboard/webapp/metrics:public_types",
         "//tensorboard/webapp/metrics:test_lib",
+        "//tensorboard/webapp/metrics:types",
         "//tensorboard/webapp/metrics/data_source:types",
         "@npm//@angular/core",
         "@npm//@ngrx/store",
